@@ -1,33 +1,48 @@
 {
   "name": "bbb-webhooks",
-  "version": "2.6.1",
+  "version": "3.0.0-alpha.0",
   "description": "A BigBlueButton mudule for events WebHooks",
+  "type": "module",
+  "scripts": {
+    "start": "node app.js",
+    "dev-start": "nodemon --watch src --ext js,json,yml,yaml --exec node app.js",
+    "test": "LOG_LEVEL=silent ALL_TESTS=true ALLOW_CONFIG_MUTATIONS=true mocha",
+    "test:webhooks": "LOG_LEVEL=silent ALL_TESTS=false ALLOW_CONFIG_MUTATIONS=true WEBHOOKS_SUITE=true mocha",
+    "test:xapi": "LOG_LEVEL=silent ALL_TESTS=false ALLOW_CONFIG_MUTATIONS=true XAPI_SUITE=true mocha",
+    "lint": "./node_modules/.bin/eslint ./",
+    "lint:file": "./node_modules/.bin/eslint",
+    "jsdoc": "./node_modules/.bin/jsdoc app.js application.js src/ -r"
+  },
   "keywords": [
     "bigbluebutton",
     "webhooks"
   ],
   "dependencies": {
-    "async": "^3.2.3",
+    "bullmq": "^4.11.4",
     "config": "^3.3.7",
-    "express": "^4.17.3",
+    "express": "^4.18.2",
     "js-yaml": "^4.1.0",
-    "lodash": "^4.17.21",
-    "redis": "^3.1.2",
-    "request": "^2.88.2",
-    "sha1": "^1.1.1",
-    "winston": "^3.7.2"
+    "luxon": "^3.4.3",
+    "node-fetch": "^3.3.2",
+    "pino": "^8.16.1",
+    "prom-client": "^14.2.0",
+    "redis": "^4.6.8",
+    "uuid": "^9.0.1"
   },
   "engines": {
-    "node": ">=12 <=20"
+    "node": ">=18"
   },
   "devDependencies": {
-    "body-parser": "^1.20.0",
+    "body-parser": "^1.20.2",
+    "eslint": "^8.49.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jsdoc": "^46.8.2",
+    "fast-xml-parser": "^4.3.2",
+    "jsdoc": "^4.0.2",
     "mocha": "^9.2.2",
-    "nock": "^13.2.4",
+    "nodemon": "^3.0.1",
+    "pino-pretty": "^10.2.3",
     "sinon": "^12.0.1",
     "supertest": "^3.4.2"
-  },
-  "scripts": {
-    "test": "mocha"
   }
 }
